Demo the format of comment.
I want to use this commit to demonstrate that how to format a git
comment.

Several rules that we should follow:
1. The first line is head of the comment. You should make it to be
more brief so that reader could understand the basic idea that why
do you want to make this fix.
2. Should make multiple lines so that reader could easier to
understand your comment especially you comment is very long.
3. It's better to have a line to have less 72 characters to easy the
read.