Demo the format of comment.
I want to use this commit to demonstrate how to format a git
comment.

Several rules that we should follow:
1. The first line is the headline of the comment. You should make it
to be more brief so that reader could understand the basic idea that why
you want to make this fix.
2. Should make the comment to be separated to multiple lines if the
comment is long than 72 characters so that reader could easily understand
your comment especially your comment is very long.
3. It's better to have a line to have less than 72 characters to make it
to be easy to read.